+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog][keepachangelog], and this project adhe
 
 ## [Unreleased]
 
+### Changed
+
+- Bump default version of JSpecify provisioned by this plugin to `1.0.1`.
+
 ## [0.7.0] - 2026-07-30
 
 ### Added